backend/modules/ground_truth/service.py
"""Ground Truth — crowdsourced field photo submissions with AI classification."""
import os
import io
import json
import uuid
import hashlib
import logging
from datetime import datetime, timezone
import httpx
from fastapi import UploadFile
from shared.db import fetch, fetchrow, execute
from shared.r2 import upload_bytes

logger = logging.getLogger(__name__)

DAMAGE_CLASSES = {0: "intact", 1: "minor damage", 2: "major damage", 3: "destroyed"}
RATE_LIMIT = 10  # submissions per IP per event per hour

async def submit_ground_report(
    event_id: str, lat: float, lon: float, 
    description: str, photo: bytes, ip_hash: str
) -> dict:
    """Process a field photo submission."""
    
    # Check rate limit
    count = await fetchrow("""
        SELECT COUNT(*) as cnt FROM ground_reports
        WHERE event_id = $1::uuid AND submitter_hash = $2
        AND created_at > now() - interval '1 hour'
    """, event_id, ip_hash)
    
    if (count["cnt"] or 0) >= RATE_LIMIT:
        return {"error": "Rate limit exceeded — max 10 submissions per hour per event"}
    
    # Validate coordinates are within 200km of event
    event = await fetchrow("SELECT lat, lon FROM events WHERE id = $1::uuid", event_id)
    if not event:
        return {"error": "Event not found"}
    
    # Compress photo to <800KB
    compressed = _compress_photo(photo)
    
    # Upload to R2
    photo_key = f"reports/{event_id}/{uuid.uuid4()}.jpg"
    photo_url = upload_bytes(photo_key, compressed, "image/jpeg")
    
    # AI classification via HuggingFace
    damage_class, confidence = await _classify_damage(compressed)
    
    # Cross-validate with satellite assessment
    satellite_class, agreement = await _cross_validate(event_id, lat, lon, damage_class)
    disputed = agreement is False
    
    # Store record
    report_id = await fetchrow("""
        INSERT INTO ground_reports 
            (event_id, damage_type, damage_class, ai_confidence, description,
             photo_storage_key, photo_url, satellite_class, agreement, disputed, submitter_hash)
        VALUES ($1::uuid, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
        RETURNING id
    """, event_id, DAMAGE_CLASSES.get(damage_class, "unknown"), damage_class, 
         confidence, description[:500] if description else None,
         photo_key, photo_url, satellite_class, agreement, disputed, ip_hash)
    
    return {
        "id": str(report_id["id"]),
        "damage_class": damage_class,
        "damage_label": DAMAGE_CLASSES.get(damage_class, "unknown"),
        "confidence": confidence,
        "disputed": disputed,
        "agreement": agreement,
        "photo_url": photo_url
    }

def _compress_photo(photo: bytes) -> bytes:
    """Compress photo to max 800KB."""
    try:
        from PIL import Image
        img = Image.open(io.BytesIO(photo))
        img = img.convert("RGB")
        img.thumbnail((800, 800))
        buf = io.BytesIO()
        quality = 85
        img.save(buf, format="JPEG", quality=quality, optimize=True)
        while buf.tell() > 800_000 and quality > 30:
            buf = io.BytesIO()
            quality -= 10
            img.save(buf, format="JPEG", quality=quality)
        return buf.getvalue()
    except Exception:
        return photo[:800_000]

async def _classify_damage(photo: bytes) -> tuple[int, float]:
    """Classify building damage via HuggingFace."""
    token = os.getenv("HUGGINGFACE_TOKEN")
    if not token:
        # Mock classification
        import random
        return random.choice([0, 1, 2, 3]), round(random.uniform(0.65, 0.95), 3)
    
    try:
        async with httpx.AsyncClient(timeout=30) as client:
            resp = await client.post(
                "https://api-inference.huggingface.co/models/microsoft/resnet-50",
                headers={"Authorization": f"Bearer {token}"},
                content=photo,
            )
            results = resp.json()
            if isinstance(results, list):
                label = results[0].get("label", "").lower()
                score = float(results[0].get("score", 0.7))
                if "destroy" in label or "collapse" in label:
                    return 3, score
                elif "major" in label or "severe" in label:
                    return 2, score
                elif "minor" in label or "damage" in label:
                    return 1, score
                return 0, score
    except Exception as e:
        logger.error(f"HuggingFace classification failed: {e}")
    
    return 1, 0.6  # Default to minor damage if classification fails

async def _cross_validate(event_id: str, lat: float, lon: float, field_class: int) -> tuple:
    """Compare field report with satellite assessment for same location."""
    analysis = await fetchrow("""
        SELECT id, damage_geojson FROM analyses 
        WHERE event_id = $1::uuid AND status IN ('complete', 'assessing_buildings', 'generating_report')
        ORDER BY created_at DESC LIMIT 1
    """, event_id)
    
    if not analysis or not analysis["damage_geojson"]:
        return None, None  # No satellite data yet
    
    geojson = analysis["damage_geojson"]
    if isinstance(geojson, str):
        geojson = json.loads(geojson)
    
    # Check if point falls within any damage polygon
    for feature in geojson.get("features", []):
        props = feature.get("properties", {})
        geom = feature.get("geometry", {})
        if geom.get("type") == "Polygon":
            coords = geom["coordinates"][0]
            if _point_in_polygon(lon, lat, coords):
                sat_class = min(props.get("severity_class", 0) // 2, 3)
                agree = abs(sat_class - field_class) <= 1
                return sat_class, agree
    
    return None, None

def _point_in_polygon(px, py, polygon) -> bool:
    """Ray casting algorithm for point-in-polygon."""
    n = len(polygon)
    inside = False
    j = n - 1
    for i in range(n):
        xi, yi = polygon[i][0], polygon[i][1]
        xj, yj = polygon[j][0], polygon[j][1]
        if ((yi > py) != (yj > py)) and (px < (xj - xi) * (py - yi) / (yj - yi) + xi):
            inside = not inside
        j = i
    return inside
