backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
httpx==0.28.1
apscheduler==3.10.4
python-dotenv==1.0.1
boto3==1.35.81
feedparser==6.0.11
numpy

python-multipart==0.0.20
reportlab==4.2.5
Pillow==11.1.0
groq==0.13.0
google-generativeai==0.8.3
huggingface-hub==0.27.0
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
